add code to index to run app from index.js
